requirements.txt
python-telegram-bot==12.2.0
pyyaml==5.1
pytz==2019.2
requests==2.22.0
